Check read DeviceID against bitstream before programming
Check that the DeviceID read from the CPLD matches an expected device ID
for the device type of the bitstream file.
Since the first read of the device ID can fail, perform two attempts to
read the device ID.
A force option -f is provided to ignore the check results and perform
the program operation anyway.